+++ hooks/src/lib/useInput.ts
@@ -1,31 +1,17 @@
 import { useState } from 'react';
-import { UseInput, ValidationResult, ValidatorProps } from './type';
+import { UseInput } from './type';
 
-const useInput = (initialValue: string, validator: ValidatorProps): UseInput => {
+const useInput = (initialValue: string): UseInput => {
   const [value, setValue] = useState(initialValue);
-  const [errorInfo, setErrorInfo] = useState<ValidationResult>({
-    isValid: true,
-    errorMessage: '',
-  });
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const validationResult = validator.onChange(event.target.value);
-    setErrorInfo(validationResult);
-    if (!validationResult.isValid) return;
     setValue(event.target.value);
-  };
-
-  const handleBlur = (event: React.FocusEvent<HTMLInputElement, Element>) => {
-    setErrorInfo(validator.onBlur(event.target.value));
   };
 
   return {
     value,
     setValue,
     handleChange,
-    handleBlur,
-    errorInfo,
-    setErrorInfo,
   };
 };
